requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
gunicorn>=21.2.0
scikit-learn>=1.4.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
numpy>=1.26.0
pypdf>=4.0.1
python-docx>=1.1.0
sentence-transformers>=2.3.1
spacy>=3.7.2
googlesearch-python>=1.2.3
google-genai>=1.0.0
httpx>=0.26.0
pbr>=5.0.0
bandit>=1.7.5
requests>=2.31.0
markupsafe>=2.1.3
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
slowapi>=0.1.9
# Dev & Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
black>=24.1.1
isort>=5.13.2
mypy>=1.8.0
structlog>=24.1.0
prometheus-fastapi-instrumentator>=6.1.0
cyclonedx-bom>=3.11.0

